#505 reorganize deal tests
